秒杀服务：秒杀功能完成

+++ gulimall-seckill/src/main/java/com/guo/gulimall/secskill/to/SecKillSkuRedisTO.java
@@ -28,11 +28,11 @@ public class SecKillSkuRedisTO {
     /**
      * 秒杀总量
      */
-    private BigDecimal secKillCount;
+    private Integer secKillCount;
     /**
      * 每人限购数量
      */
-    private BigDecimal secKillLimit;
+    private Integer secKillLimit;
     /**
      * 排序
      */
